+++ src/drawMode1.h
@@ -3,5 +3,5 @@
 class DM1: public Fractal{
     public:
         DM1(bool);
-        void draw(int,int,int,int);
+        void draw(float,float,float,int);
 };
